Header file Group.hpp moved to Group/Group.hpp
