+++ opm/core/wells/WellsManager.hpp
@@ -26,7 +26,7 @@
 
 #include <opm/core/wells/WellCollection.hpp>
 #include <opm/core/wells/WellsGroup.hpp>
-#include <opm/parser/eclipse/EclipseState/Schedule/GroupTree.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Group/GroupTree.hpp>
 
 #include <opm/grid/utility/CompressedPropertyAccess.hpp>
